requirements.txt
annotated-types==0.7.0
anyio==4.4.0
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
exceptiongroup==1.2.2
fastapi==0.112.1
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
moviepy==1.0.3
msgpack==1.0.8
numba==0.60.0
numpy==2.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
piexif==1.1.3
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyWavelets==1.7.0
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.4.0
starlette==0.38.2
threadpoolctl==3.5.0
tqdm==4.66.5
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
